+++ packages/ricos-content/web/src/RicosContentAPI/builder-utils.ts
@@ -1,7 +1,7 @@
 import { isNumber, isString, isArray, isObject } from 'lodash';
-import { flow } from 'fp-ts/lib/function';
-import * as A from 'fp-ts/lib/Array';
-import * as O from 'fp-ts/lib/Option';
+import { flow } from 'fp-ts/function';
+import * as A from 'fp-ts/Array';
+import * as O from 'fp-ts/Option';
 import { ParagraphData, RichContent, TextData, Node } from 'ricos-schema';
 import { resolveFirstRight } from '../fp-utils';
 import { ListItemData } from '../types';