+++ CHANGELOG.md
@@ -20,6 +20,11 @@
 </details>
 <hr/>
 
+## 8.47.10  (Jul 27, 2021)
+### :bug: Bug Fix
+  - `ricos-content`
+    - [#2784](https://github.com/wix/ricos/pull/2784) fromHtml: ckeditor parser rules fixed
+
 ## 8.47.9 (July 26, 2021)
 
 ### :house: Internal